+++ bsp_cc26x2/i2c/i2c_commands.c
@@ -20,7 +20,7 @@ bool i2c_write_command(int32_t argc, char* argv[]) {
         uint8_t i2c_num = 0;
         uint8_t i2c_addr = 0;
         uint8_t array[256];
-        uint16_t array_len = 0;
+        uint32_t array_len = 0;
         if(true == res) {
             res = try_str2uint8(argv[0], &i2c_num);
             if(false == res) {
