+++ protocols/ubx_protocol/ublox_command.c
@@ -108,7 +108,7 @@ bool ubx_send_command(int32_t argc, char* argv[]) {
     bool res = false;
     uint8_t class_num = 0;
     uint8_t id = 0;
-    uint16_t payload_len = 0U;
+    uint32_t payload_len = 0U;
     uint8_t payload[100];
     if(2 <= argc) {
         res = true;
